refactored whitelist processing

+++ google_cloud_run_services/server.py
@@ -255,6 +255,17 @@ def run_sql(conn, sql_query, *params):
 # Query to add ip to the restricted_ips table
 #run_sql("""INSERT INTO restricted_ips (ip) VALUES ('210.3.222.157')""")
 
+def is_user_on_whitelist(conn, user_ip):
+    """Check if the user is on the whitelist"""
+    if conn is None or not user_ip:
+        return False
+
+    if not re.match(r"^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$", user_ip):
+        return False
+
+    rows = run_sql(conn, "SELECT COUNT(ip) FROM whitelist_ips WHERE ip=%s", (user_ip,))
+    return rows and int(rows[0][0]) > 0
+
 def exceeds_rate_limit(conn, user_ip, params):
     """Rate limit requests based on user ip address"""
 
@@ -263,10 +274,12 @@ def exceeds_rate_limit(conn, user_ip, params):
     #SELECT ip, count(*) FROM log WHERE event_name like '%computed' AND duration > 2 AND logtime >= NOW() - INTERVAL '20 minutes' GROUP BY ip ORDER BY count DESC;
     #"""
 
-    if conn is None:
-        return False
-
     try:
+        if conn is None:
+            return False
+
+        if is_user_on_whitelist(conn, params.get("ip")):
+            return False
 
         # check if the user has exceeded the rate limit or is on the list of restricted IPs
         rows = run_sql(conn, "SELECT COUNT(ip) FROM restricted_ips WHERE ip=%s AND created >= NOW() - INTERVAL '1 weeks'", (user_ip,))
@@ -276,33 +289,27 @@ def exceeds_rate_limit(conn, user_ip, params):
 
         rows = run_sql(conn, "SELECT COUNT(ip) FROM log WHERE event_name LIKE %s AND ip=%s AND logtime >= NOW() - INTERVAL '7 minutes'", ("%computed%", user_ip))
         did_user_exceed_rate_limit = rows and int(rows[0][0]) >= 50
-        if did_user_exceed_rate_limit:
-            ip_for_whitelist_check = params.get("ip")
-            if not ip_for_whitelist_check or not re.match(r"^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$", ip_for_whitelist_check):
-                ip_for_whitelist_check = user_ip
-            rows = run_sql(conn, "SELECT COUNT(ip) FROM whitelist_ips WHERE ip=%s", (ip_for_whitelist_check,))
-            is_user_on_whitelist = rows and int(rows[0][0]) > 0
-            if not is_user_on_whitelist:
-                # the user has exceeded the rate limit: computing scores for 50 or more variants in the last 7 minutes
-                rows = run_sql(conn, "SELECT COUNT(ip) FROM log WHERE event_name='rate_limit_exceeded' AND ip=%s AND logtime >= NOW() - INTERVAL '5 minutes'", (user_ip,))
-                user_hit_rate_limit_exceeded_recently = rows and int(rows[0][0]) > 0
-                if not user_hit_rate_limit_exceeded_recently:
-                    # the user will receive at most one "rate_limit_exceeded" event every 5 minutes
-                    log(conn, f"rate_limit_exceeded", ip=user_ip)
-                    rows = run_sql(conn, "SELECT COUNT(ip) FROM log WHERE event_name='rate_limit_exceeded' AND ip=%s AND logtime >= NOW() - INTERVAL '1 days'", (user_ip,))
-                    user_triggered_too_many_rate_limit_exceeded_errors_today = rows and int(rows[0][0]) >= 5
-                    if user_triggered_too_many_rate_limit_exceeded_errors_today:
-                        # the user has hit the limit of 5 or more "rate_limit_exceeded" events during the last 24 hours
-                        rows = run_sql(conn, "SELECT COUNT(ip) FROM restricted_ips WHERE ip=%s", (user_ip,))
-                        need_to_delete_previous_restricted_ip_record = rows and int(rows[0][0]) > 0
-                        if need_to_delete_previous_restricted_ip_record:
-                            # delete the previous record
-                            run_sql(conn, "DELETE FROM restricted_ips WHERE ip=%s", (user_ip,))
+        if did_user_exceed_rate_limit and not is_user_on_whitelist(conn, user_ip):
+            # the user has exceeded the rate limit: computing scores for 50 or more variants in the last 7 minutes
+            rows = run_sql(conn, "SELECT COUNT(ip) FROM log WHERE event_name='rate_limit_exceeded' AND ip=%s AND logtime >= NOW() - INTERVAL '5 minutes'", (user_ip,))
+            user_hit_rate_limit_exceeded_recently = rows and int(rows[0][0]) > 0
+            if not user_hit_rate_limit_exceeded_recently:
+                # the user will receive at most one "rate_limit_exceeded" event every 5 minutes
+                log(conn, f"rate_limit_exceeded", ip=user_ip)
+                rows = run_sql(conn, "SELECT COUNT(ip) FROM log WHERE event_name='rate_limit_exceeded' AND ip=%s AND logtime >= NOW() - INTERVAL '1 days'", (user_ip,))
+                user_triggered_too_many_rate_limit_exceeded_errors_today = rows and int(rows[0][0]) >= 5
+                if user_triggered_too_many_rate_limit_exceeded_errors_today:
+                    # the user has hit the limit of 5 or more "rate_limit_exceeded" events during the last 24 hours
+                    rows = run_sql(conn, "SELECT COUNT(ip) FROM restricted_ips WHERE ip=%s", (user_ip,))
+                    need_to_delete_previous_restricted_ip_record = rows and int(rows[0][0]) > 0
+                    if need_to_delete_previous_restricted_ip_record:
+                        # delete the previous record
+                        run_sql(conn, "DELETE FROM restricted_ips WHERE ip=%s", (user_ip,))
 
-                        # block the user's IP for 1 week
-                        run_sql(conn, "INSERT INTO restricted_ips (ip) VALUES (%s)", (user_ip,))
+                    # block the user's IP for 1 week
+                    run_sql(conn, "INSERT INTO restricted_ips (ip) VALUES (%s)", (user_ip,))
 
-                return RATE_LIMIT_ERROR_MESSAGE
+            return RATE_LIMIT_ERROR_MESSAGE
 
     except Exception as e:
         print(f"Error while checking rate limit: {e}", flush=True)
